运营中枢/scripts/aiye_im_bridge/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
pyyaml>=6.0
websockets>=13.0
